snapshot of project "lynx", label v2-8-7dev_12b

+++ WWW/Library/Implementation/HTPlain.c
@@ -1,5 +1,5 @@
 /*
- * $LynxId: HTPlain.c,v 1.42 2008/09/06 14:45:09 tom Exp $
+ * $LynxId: HTPlain.c,v 1.44 2009/01/03 01:23:21 tom Exp $
  *
  *		Plain text object		HTWrite.c
  *		=================
@@ -159,13 +159,13 @@ static void HTPlain_put_character(HTStream *me, char c)
 	    UCode_t value = (UCode_t) FROMASCII((TOASCII(UCH(c)) - 160));
 
 	    name = HTMLGetEntityName(value);
-	    len = strlen(name);
-	    for (low = 0, high = HTML_dtd.number_of_entities;
+	    len = (int) strlen(name);
+	    for (low = 0, high = (int) HTML_dtd.number_of_entities;
 		 high > low;
 		 diff < 0 ? (low = i + 1) : (high = i)) {
 		/* Binary search */
 		i = (low + (high - low) / 2);
-		diff = AS_ncmp(HTML_dtd.entity_names[i], name, len);
+		diff = AS_ncmp(HTML_dtd.entity_names[i], name, (unsigned) len);
 		if (diff == 0) {
 		    HText_appendText(me->text,
 				     LYCharSets[me->outUCLYhndl][i]);
@@ -572,7 +572,7 @@ static void HTPlain_write(HTStream *me, const char *s, int l)
 		 * Out of luck, so use the UHHH notation (ugh).  - gil
 		 */
 		/* S/390 -- gil -- 0517 */
-		sprintf(replace_buf, "U%.2lX", TOASCII(code));
+		sprintf(replace_buf, "U%.2lX", (unsigned long) TOASCII(code));
 		HText_appendText(me->text, replace_buf);
 	    }
 #ifdef NOTDEFINED
